Move packages to internal/ because they shouldn't be imported
I tend to ignore doing this because I am lazy/remember the times before
internal. But it does seem to be the way things should be done, and does neaten
up the root of the repo.

I don't think moving main.go to cmd/opg-file-service/ would make sense as there
will only ever be the 1 main package.